Smashrun: fetch activities without pagination when doing exhaustive sync
Chris Lukic suggested that we disable pagination when performing an
exhaustive sync to help reduce the load on his servers.

While we're at it, we can also take advantage of the new(-ish) limit
kwarg to fix an off-by-one error where we always fetched 30 activities
when we only intended to fetch 20.

Fixes #448